Chapter26-PropertyWrappers-GoldChallenge - Complited p. 678

+++ PercentageClamping.playground/Contents.swift
@@ -2,7 +2,7 @@ import Cocoa
 
 struct Car {
     @Percentge var fuelLevel: Double = 1.0
-    @Percentge var wiperFluidLevel: Double = 0.5
+    @Percentge var wiperFluidLevel: Float = 0.5
     @Percentge(upperBound: 2.0) var stereoVolume: Double = 1.0
 }
 
@@ -16,6 +16,6 @@ var myCar = Car()
 print(myCar)
 myCar.fuelLevel = 1.1
 print("Fuel:", myCar.fuelLevel)
-myCar.stereoVolume = 2.5 
+myCar.stereoVolume = 2.5
 print("Volume:", myCar.stereoVolume)
 print("Prijected volume:", myCar.$stereoVolume)
